Convert number-like cell values into numbers
It means now if there's a cell value in the HTML table like "$12,345.67", it'll be appear in the CSV as 12345.67. Numeric sorting, multiplication, etc is now possible straight from the output.

It can be switched off using command-line parameters, and group/decimal/currency symbols can be set in the same way.