Accept one table element or a list of tr elements
Since we're not interested in the "table" element per se, but the
content of its "tr" elements, there's no good reason not to accept a CSS
selector or XPath expression that matches a list of "tr" elements. This
has the benefit of allowing table rows to be filtered — for example
removing junk header rows, or rows with a certain class attribute.