Simplify expression argument and related options
The CLI now doesn't allow the user to explicitly specify whether the
expression is an index, CSS selector, or XPath expression. Instead it
tries those types one-by-one (in that order), using the first type that
doesn't cause an error. This means if the expression is an integer it's
used as an index, if it's a valid CSS selector it's used as such, else
if it's a valid XPath expression it's used like that.

The result is that the user doesn't need to fiddle with command-line
switches to use CSS or XPath.