Expression now passed as option not argument
Trying to have two optional arguments (expression and filename/URL)
meant it was ambiguous when only one was given. For example:

    htmltab table.html

Does this mean find the first table in the file table.html, or does it
mean find the table with the class 'html' from stdin? The solution is to
pass the expression as an option (--select) and leave the filename/URL
as an optional argument.