Ensure all rows have the same number of columns
CSV files tend to have the same number of columns/fields/cells in each
row — indeed, it's demanded by RFC 4180, the closest thing to a CSV
spec. But HTML tables have no such limitation and rows can be "ragged",
containing rows with differing number of cells.

This can cause problems with tools: for example, if the first row has
three columns but the rest of the rows have five, CSVKit will only
recognise three columns on each row.

To solve this problem, empty cells are added to rows that are shorter
than the longest row, until all rows are the same length.