Switch from namespace package to standard package
Fixes #6. In commit 9637a032 I moved the project to a src layout, and in
the process I switched to using a namespace package (that is, I removed
the __init__.py from the htmltab package). While Pip could install the
package without a problem, Pipx seems not to like it.

There's no need for htmltab to be a namespace package (other than the
repo being cleaner without the __init__.py), so I've added __init__.py
back in.

https://packaging.python.org/en/latest/guides/packaging-namespace-packages/